fix: Move shareDining funtion to ui
Move shareDining funtion to ui
